move tests.py
