+++ tests/e2e/cli/test_cli.py
@@ -7,7 +7,7 @@ import os
 
 import pytest
 
-from otx.cli.utils.tests import (
+from tests.test_suite.run_test_command import (
     otx_build_auto_config,
     otx_build_backbone_testing,
     otx_find_testing,
