+++ tests/integration/cli/detection/test_instance_segmentation.py
@@ -10,7 +10,7 @@ import torch
 
 from otx.api.entities.model_template import parse_model_template
 from otx.cli.registry import Registry
-from otx.cli.utils.tests import (
+from tests.test_suite.run_test_command import (
     get_template_dir,
     nncf_optimize_testing,
     otx_deploy_openvino_testing,
