+++ tests/run_model_templates_tests.py
@@ -4,7 +4,7 @@ import os
 import sys
 from subprocess import run
 
-from ote_cli.utils.tests import collect_env_vars
+from tests.test_suite.run_test_command import collect_env_vars
 
 ALGO_ROOT_DIR = "external"
 ALGO_DIRS = [
